fixed compiler warnings

+++ src/goryachev/memsafecrypto/Pack.java
@@ -93,7 +93,7 @@ public abstract class Pack
 	{
 		int hi = bigEndianToInt(bs, off);
 		int lo = bigEndianToInt(bs, off + 4);
-		return ((long)(hi & 0xffffffffL) << 32) | (long)(lo & 0xffffffffL);
+		return ((hi & 0xffffffffL) << 32) | (lo & 0xffffffffL);
 	}
 
 
@@ -307,7 +307,7 @@ public abstract class Pack
 	{
 		int lo = littleEndianToInt(bs, off);
 		int hi = littleEndianToInt(bs, off + 4);
-		return ((long)(hi & 0xffffffffL) << 32) | (long)(lo & 0xffffffffL);
+		return ((hi & 0xffffffffL) << 32) | (lo & 0xffffffffL);
 	}
